requirements.txt
django-braces